MINOR: Move some integration tests to unit tests
